ci(docs): configure the GH Actions workflow to build and deploy the docs (#37)
This PR includes a new GitHub Actions workflow to automate the building
and deployment procedure of the project documentations.

The workflow utilises some Actions (officially provided by the Astro
devs) to build and generate the documentations. After the static assets
are generated and built, the documentations is then deployed to GitHub
Pages (using another officially provided GitHub Actions from the GitHub
devs).