feat(parser): update the parser to match "breaking changes" (#49)
This PR includes some changes to the parser so that it can match for an
optional `!` mark in the commit message. The exclamation mark signifies
a breaking change introduced in a particular commit message. In addition
to the parsing logic, the PR also includes relevant tests for the logic
t be tested on.

Closes #48.