feat(parser): update the parser to parse the footer (#54)
This PR updates the parser to be able to properly parse the commit
message when a footer is also provided. The full list of updates
introduced with the PR are as follows:

1. Refactored the `parser` package with new structs (and receiver
methods) along with standalone functions to parse the footer properly.
2. Check the footer for special keywords like `BREAKING CHANGES` to
introduce a breaking change introduced by a commit and so on.
3. Implemented functions to parse the entire commit message into
sections like the header, description and the footer for improved
validation down the pipeline.
4. Introduced more tests for robustness and developement confidence into
the implementation.

Closes #52.